# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Typeform Clone API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/creator/forms": {
      "get": {
        "tags": [
          "Creator"
        ],
        "summary": "List Forms",
        "operationId": "list_forms_api_v1_creator_forms_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Creator"
        ],
        "summary": "Create Form",
        "operationId": "create_form_api_v1_creator_forms_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateForm"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/creator/forms/{form_id}": {
      "get": {
        "tags": [
          "Creator"
        ],
        "summary": "Get Form",
        "operationId": "get_form_api_v1_creator_forms__form_id__get",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Creator"
        ],
        "summary": "Rename Form",
        "operationId": "rename_form_api_v1_creator_forms__form_id__patch",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          },
          {
            "name": "if-match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Creator"
        ],
        "summary": "Delete Form",
        "operationId": "delete_form_api_v1_creator_forms__form_id__delete",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/creator/forms/{form_id}/draft": {
      "put": {
        "tags": [
          "Creator"
        ],
        "summary": "Save Draft",
        "operationId": "save_draft_api_v1_creator_forms__form_id__draft_put",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          },
          {
            "name": "if-match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDraft"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/creator/forms/{form_id}/duplicate": {
      "post": {
        "tags": [
          "Creator"
        ],
        "summary": "Duplicate Form",
        "operationId": "duplicate_form_api_v1_creator_forms__form_id__duplicate_post",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          },
          {
            "name": "if-match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/creator/forms/{form_id}/publish": {
      "post": {
        "tags": [
          "Creator"
        ],
        "summary": "Publish Form",
        "operationId": "publish_form_api_v1_creator_forms__form_id__publish_post",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          },
          {
            "name": "if-match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/creator/forms/{form_id}/unpublish": {
      "post": {
        "tags": [
          "Creator"
        ],
        "summary": "Unpublish Form",
        "operationId": "unpublish_form_api_v1_creator_forms__form_id__unpublish_post",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormMetadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/creator/forms/{form_id}/versions": {
      "get": {
        "tags": [
          "Creator"
        ],
        "summary": "Versions",
        "operationId": "versions_api_v1_creator_forms__form_id__versions_get",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/creator/forms/{form_id}/responses": {
      "get": {
        "tags": [
          "Creator"
        ],
        "summary": "Responses",
        "operationId": "responses_api_v1_creator_forms__form_id__responses_get",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "version_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/creator/forms/{form_id}/responses/{response_id}": {
      "get": {
        "tags": [
          "Creator"
        ],
        "summary": "Response Detail",
        "operationId": "response_detail_api_v1_creator_forms__form_id__responses__response_id__get",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          },
          {
            "name": "response_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Response Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/creator/forms/{form_id}/summary": {
      "get": {
        "tags": [
          "Creator"
        ],
        "summary": "Summary",
        "operationId": "summary_api_v1_creator_forms__form_id__summary_get",
        "parameters": [
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          },
          {
            "name": "version_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateForm": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title",
            "default": "Untitled form"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CreateForm"
      },
...
      "DraftDefinition": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionDefinition"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Questions"
          },
          "theme_settings": {
            "$ref": "#/components/schemas/ThemeSettings"
          },
          "ending_settings": {
            "$ref": "#/components/schemas/EndingSettings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "DraftDefinition"
      },
...
      "FormDetail": {
        "properties": {
          "form": {
            "$ref": "#/components/schemas/FormMetadata"
          },
          "draft": {
            "$ref": "#/components/schemas/DraftDefinition"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "form",
          "draft"
        ],
        "title": "FormDetail"
      },
      "FormList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FormMetadata"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "FormList"
      },
      "FormMetadata": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published"
          },
          "public_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Slug"
          },
          "draft_revision": {
            "type": "integer",
            "title": "Draft Revision"
          },
          "last_save_mutation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Save Mutation Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "response_count": {
            "type": "integer",
            "title": "Response Count"
          },
          "seed_response_count": {
            "type": "integer",
            "title": "Seed Response Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "is_published",
          "public_slug",
          "draft_revision",
          "last_save_mutation_id",
          "created_at",
          "updated_at",
          "response_count",
          "seed_response_count"
        ],
        "title": "FormMetadata"
      },
      "FormSummary": {
        "properties": {
          "version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Id"
          },
          "version_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Number"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "response_count": {
            "type": "integer",
            "title": "Response Count"
          },
          "seed_response_count": {
            "type": "integer",
            "title": "Seed Response Count"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionSummary"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version_id",
          "version_number",
          "title",
          "response_count",
          "seed_response_count",
          "questions"
        ],
        "title": "FormSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperationResult": {
        "properties": {
          "status": {
            "type": "string",
            "const": "deleted",
            "title": "Status",
            "default": "deleted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "OperationResult"
      },
...
      "RenameForm": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "RenameForm"
      },
      "ResponseDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version_id": {
            "type": "string",
            "title": "Version Id"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "version_number": {
            "type": "integer",
            "title": "Version Number"
          },
          "form_title": {
            "type": "string",
            "title": "Form Title"
          },
          "is_seed": {
            "type": "boolean",
            "title": "Is Seed"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AnswerView"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "version_id",
          "submitted_at",
          "version_number",
          "form_title",
          "is_seed",
          "answers"
        ],
        "title": "ResponseDetail"
      },
...
      "ResponseList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResponseItem"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ResponseList"
      },
      "SaveDraft": {
        "properties": {
          "mutation_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}(-[0-9a-f]{4}){3}-[0-9a-f]{12}$",
            "title": "Mutation Id"
          },
          "definition": {
            "$ref": "#/components/schemas/DraftDefinition"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mutation_id",
          "definition"
        ],
        "title": "SaveDraft"
      },
...
      "ThemeSettings": {
        "properties": {
          "font": {
            "type": "string",
            "const": "inter",
            "title": "Font",
            "default": "inter"
          },
          "background": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Background",
            "default": "#ffffff"
          },
          "text": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Text",
            "default": "#2a222b"
          },
          "accent": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Accent",
            "default": "#2563eb"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ThemeSettings"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VersionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VersionInfo"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "VersionList"
      }
    }
  }
}